Before Nodemailer function was inside of sseController.js. No I have moved it to a seperat file named email.js. No only the function name is called inside the sseController named  sendEmailWithNodemailer(booking).   Its tested an working just fine.
