+++ server/src/controller/sseController.js
@@ -3,15 +3,11 @@ import { fetchCollection } from "../mongo/mongoClient.js";
 import jwtUtil from "../util/jwtUtil.js";
 import idUtil from "../util/idUtil.js";
 import calcTotalPrice from "../util/calcTotalPrice.js";
-import nodemailer from 'nodemailer';
 import * as dotenv from "dotenv";
+
 dotenv.config();
-import {dirname, join as pathJoin} from "path";
-import { fileURLToPath } from "url";
 
-const __dirname = dirname(fileURLToPath(import.meta.url));
-const logoPath = pathJoin(__dirname, "..", "assets"  );
-
+import sendEmailWithNodemailer from "../email.js";
 
 let clients = [];
 
@@ -61,51 +57,10 @@ const postBooking = async (req, res) => {
       price: totalPrice,
       status: true 
     }
-
-    // Nodemailer börjar här
-    const transporter = nodemailer.createTransport({
-      host: process.env.host,
-      port: 587,
-      secure: false,
-      auth: {
-        user: process.env.email,
-        pass: process.env.emailPassword, // Om du använder tvåfaktorsautentisering (2FA), använd ett appspecifikt lösenord här
-      },
-      tls: {
-          rejectUnauthorized: false,
-      }
-    });
-
-
-const mailOptions = {
-  from: `"Filmvisarna 🎥🍿 ${process.env.email}` ,
-  to: booking.customerEmail, 
-  subject: 'Bokningsbekräftelse',
-  text:`  `,
-  html: `  <div style="border:#DACA88; border-width:2px; border-style:solid; padding:10px; text-align:center; width:400px; border-radius:8px; font-size:16px;">
-  <h2 style="color:black;">Tack för din bokning.</h2> 
-  <p>Ditt bokningsnummer är: <span style="font-weight:800">${booking.bookingId}</span> 
-  <br><h1></h1> 
-  Vi på Filmvisarna önskar en underbar biostund.
-  Bokningsnummret visar du upp i kassan i samband <br> med betalning.
-  <br>
-  <br>
-  Välkommen!</p> 
-  <br><img width="40px" src="cid:${process.env.email}">
-  <br>
-  </div>`,
-  attachments: [
-    {   // utf-8 string as an attachment
-      filename: 'logo.png',
-        path: `${logoPath}/logo.png`,
-        cid: process.env.email //same cid value as in the html img src
-    }
-  ]
-};
-
-  transporter.sendMail(mailOptions)
-
-  // Nodemailer slutar ---------------------------------------------------------------
+   
+    // Sending emailconfirmation from email.js
+    sendEmailWithNodemailer(booking)
+   
 
     await fetchCollection("bookings").insertOne(booking)
